Update README (#57)
I don't know why, but the roundness adjustment feature(#56) is applied
in [Preview](https://avatar-jig9j3gwu.vercel.sh/rauchg?rounded=60), but
not yet in the [Production](https://avatar.vercel.sh/rauchg?rounded=60)
page.

If it is applied, I would like to add the roundness adjustment feature
to the README for users.

The part marked as `<!-- Delete This -->` is the part that will be
deleted when roundness adjustment feature is applied to the Production
page.

---------

Co-authored-by: Steven <steven@ceriously.com>